+++ ebustoolbox/models.py
@@ -36,6 +36,7 @@ class EnumScenarioType(models.TextChoices):
     SOURCE = "SOURCE"
     MUTATION = "MUTATION"
     SIMULATION = "SIMULATION"
+    PUBLIC_DATA = "PUBLIC_DATA"
 
 
 class EnumSimulationType(models.TextChoices):
@@ -88,15 +89,15 @@ class Scenario(models.Model):
     name_short = models.TextField(blank=True, null=True)
     parent = models.ForeignKey("self", on_delete=models.CASCADE, null=True, blank=True)
 
-    scenario_type = models.CharField(choices=EnumScenarioType.choices, null=True)
-    description = models.TextField(blank=True, null=True)
+    scenario_type = models.CharField(choices=EnumScenarioType.choices, null=True, blank=True)
+    description = models.TextField(null=True, blank=True)
 
     # Set to now() on the database side
     created = models.DateTimeField(auto_now_add=True, db_default=Now())
     task_id = models.UUIDField(default=None, null=False, unique=True)
     finished = models.DateTimeField(default=None, null=True, blank=True)
-    simba_options = models.JSONField(default=dict, null=True)
-    eflips_depot_options = models.JSONField(default=dict, null=True)
+    simba_options = models.JSONField(default=dict, null=True, blank=True)
+    eflips_depot_options = models.JSONField(default=dict, null=True, blank=True)
     tco_parameters = models.JSONField(
         default=dict,
         null=True,
@@ -117,6 +118,7 @@ class Scenario(models.Model):
             "pef_energy": 0.038,
             "pef_insurance": 0.02,
         },
+        blank=True,
     )
 
     manager = models.ForeignKey(
@@ -223,6 +225,7 @@ class BatteryType(models.Model):
     tco_parameters = models.JSONField(
         default=dict,
         null=True,
+        blank=True,
         db_default={"useful_life": 7, "procurement_cost": None, "cost_escalation": 0.01},
     )
 
@@ -730,11 +733,13 @@ class Rotation(models.Model):
 
     scenario = models.ForeignKey(Scenario, null=False, on_delete=models.CASCADE)
 
-    name = models.TextField(blank=False, null=True)
+    name = models.TextField(blank=True, null=True)
     vehicle_type = models.ForeignKey(VehicleType, null=False, blank=True, on_delete=models.CASCADE)
 
     # SimBA specific data to make SimBA simulations reproducible
-    vehicle = models.ForeignKey(Vehicle, on_delete=models.SET_DEFAULT, default=None, null=True)
+    vehicle = models.ForeignKey(
+        Vehicle, on_delete=models.SET_DEFAULT, default=None, null=True, blank=True
+    )
     allow_opportunity_charging = models.BooleanField(default=True, null=False)
 
     def get_distance(self):
@@ -1083,7 +1088,7 @@ class Station(models.Model):
         db_table = "Station"
 
     # Map Engine models need geom and name as first columns
-    geom = models.PointField(dim=3, srid=4326, null=True)  # without z elevation
+    geom = models.PointField(dim=3, srid=4326, null=True, blank=True)  # without z elevation
     name = models.TextField(null=False)
     name_short = models.TextField(null=True, blank=True)
     scenario = models.ForeignKey(Scenario, null=False, on_delete=models.CASCADE)
@@ -1091,22 +1096,25 @@ class Station(models.Model):
     is_electrified = models.BooleanField(default=False)
     is_electrifiable = models.BooleanField(default=False)
     charge_type = models.CharField(
-        max_length=4, choices=EnumChargeType.choices, null=True, default=None
+        max_length=4, choices=EnumChargeType.choices, null=True, default=None, blank=True
     )
     voltage_level = models.CharField(
-        max_length=5, choices=EnumVoltageLevel.choices, null=True, default=None
+        max_length=5, choices=EnumVoltageLevel.choices, null=True, default=None, blank=True
     )
-    amount_charging_places = models.IntegerField(default=0, null=True)
-    power_per_charger = models.FloatField(default=None, null=True)
-    power_total = models.FloatField(default=None, null=True)
+    amount_charging_places = models.IntegerField(default=0, null=True, blank=True)
+    power_per_charger = models.FloatField(default=None, null=True, blank=True)
+    power_total = models.FloatField(default=None, null=True, blank=True)
 
     tco_parameters = models.JSONField(
         default=dict,
         null=True,
         db_default={"useful_life": 20, "procurement_cost": None, "cost_escalation": 0.02},
+        blank=True,
     )
 
-    charging_point_type = models.ForeignKey(ChargingPointType, null=True, on_delete=models.CASCADE)
+    charging_point_type = models.ForeignKey(
+        ChargingPointType, null=True, blank=True, on_delete=models.CASCADE
+    )
 
     stations = models.ManyToManyField("Route", through="AssocRouteStation")
     """Stations along this route. Ordered by `elapsed_distance`."""
@@ -1307,14 +1315,14 @@ class Route(models.Model):
         # do this in Django though.
 
     # Shape of the route with height data
-    geom = models.LineStringField(dim=3, srid=4326, null=True)
+    geom = models.LineStringField(dim=3, srid=4326, null=True, blank=True)
     distance = models.FloatField(default=None, null=False)
 
     objects = FastUpdateManager()
     name = models.TextField(default=None, null=False, blank=True)
     name_short = models.TextField(default=None, null=True, blank=True)
     scenario = models.ForeignKey(Scenario, null=False, on_delete=models.CASCADE)
-    line = models.ForeignKey(Line, null=True, on_delete=models.CASCADE)
+    line = models.ForeignKey(Line, null=True, blank=True, on_delete=models.CASCADE)
     headsign = models.TextField(default=None, null=True, blank=True)
 
     departure_station = models.ForeignKey(
@@ -1352,7 +1360,7 @@ class AssocRouteStation(models.Model):
     elapsed_distance = models.FloatField(null=False)
     """The distance in m that the bus has traveled when it reached this stop."""
 
-    location = models.PointField(dim=3, srid=4326, null=True)
+    location = models.PointField(dim=3, srid=4326, null=True, blank=True)
     """An optional precise location of the this route's stop at the station. Use WGS84 coordinates (EPSG:4326)."""
 
 
@@ -1428,7 +1436,7 @@ class Trip(models.Model):
         max_length=9, choices=EnumTripType.choices, default=EnumTripType.PASSENGER_TRIP
     )
 
-    loaded_mass = models.FloatField(default=None, null=True)
+    loaded_mass = models.FloatField(default=None, null=True, blank=True)
 
     # If time resolution is minutes, there might be trips with 0 minutes duration. To resolve
     # division by 0, we use a minimal duration of 60 seconds
@@ -1601,12 +1609,12 @@ class Event(models.Model):
     objects = FastUpdateManager()
     scenario = models.ForeignKey(Scenario, null=False, on_delete=models.CASCADE)
     vehicle_type = models.ForeignKey(VehicleType, null=False, on_delete=models.CASCADE)
-    vehicle = models.ForeignKey(Vehicle, null=True, on_delete=models.CASCADE)
+    vehicle = models.ForeignKey(Vehicle, null=True, blank=True, on_delete=models.CASCADE)
 
     #
-    station = models.ForeignKey(Station, null=True, on_delete=models.CASCADE)
-    trip = models.ForeignKey(Trip, null=True, on_delete=models.CASCADE)
-    area = models.ForeignKey("Area", null=True, on_delete=models.CASCADE)
+    station = models.ForeignKey(Station, null=True, blank=True, on_delete=models.CASCADE)
+    trip = models.ForeignKey(Trip, null=True, blank=True, on_delete=models.CASCADE)
+    area = models.ForeignKey("Area", null=True, blank=True, on_delete=models.CASCADE)
 
     subloc_no = models.IntegerField(null=True, blank=True)
     time_start = models.DateTimeField()
@@ -1620,7 +1628,7 @@ class Event(models.Model):
     )
     description = models.TextField(null=True, blank=True)
 
-    timeseries = models.JSONField(default=dict, null=True)
+    timeseries = models.JSONField(default=dict, null=True, blank=True)
 
     def save(self, *args, **kwargs):
         # Exactly one of the following has to be non-null
@@ -1700,7 +1708,7 @@ class Depot(models.Model):
     scenario = models.ForeignKey(Scenario, null=False, on_delete=models.CASCADE)
     name = models.TextField(null=False, blank=False)
     name_short = models.TextField(null=True, blank=True)
-    bounding_box = models.PolygonField(dim=2, srid=4326, null=True, default=None)
+    bounding_box = models.PolygonField(dim=2, srid=4326, null=True, blank=True, default=None)
     station = models.ForeignKey(Station, null=False, on_delete=models.CASCADE)  # Added in schema v3
 
     default_plan = models.OneToOneField("Plan", null=False, on_delete=models.CASCADE)
@@ -1729,11 +1737,11 @@ class Process(models.Model):
 
     scenario = models.ForeignKey(Scenario, on_delete=models.CASCADE)
     name = models.TextField(null=False)
-    name_short = models.TextField(null=True)
-    duration = models.DurationField(null=True)
-    electric_power = models.FloatField(null=True)
+    name_short = models.TextField(null=True, blank=True)
+    duration = models.DurationField(null=True, blank=True)
+    electric_power = models.FloatField(null=True, blank=True)
     dispatchable = models.BooleanField(null=False)
-    availability = models.JSONField(default=dict, null=True)
+    availability = models.JSONField(default=dict, null=True, blank=True)
     plans = models.ManyToManyField(Plan, through="AssocPlanProcess")
 
 
@@ -1757,15 +1765,19 @@ class Area(models.Model):
 
     scenario = models.ForeignKey(Scenario, null=False, on_delete=models.CASCADE)
     depot = models.ForeignKey(Depot, null=False, on_delete=models.CASCADE)
-    vehicle_type = models.ForeignKey(VehicleType, null=True, on_delete=models.CASCADE)
-    name = models.TextField(null=True)
-    name_short = models.TextField(null=True)
-    area_type = models.CharField(max_length=14, choices=AreaType.choices, null=True, default=None)
-    bounding_box = models.PolygonField(dim=2, srid=4326, null=True, default=None)
-    row_count = models.IntegerField(null=True, default=None)
+    vehicle_type = models.ForeignKey(VehicleType, null=True, blank=True, on_delete=models.CASCADE)
+    name = models.TextField(null=True, blank=True)
+    name_short = models.TextField(null=True, blank=True)
+    area_type = models.CharField(
+        max_length=14, choices=AreaType.choices, null=True, blank=True, default=None
+    )
+    bounding_box = models.PolygonField(dim=2, srid=4326, null=True, blank=True, default=None)
+    row_count = models.IntegerField(null=True, blank=True, default=None)
     capacity = models.IntegerField(null=False)
     processes = models.ManyToManyField(Process, through="AssocAreaProcess")
-    charging_point_type = models.ForeignKey(ChargingPointType, null=True, on_delete=models.CASCADE)
+    charging_point_type = models.ForeignKey(
+        ChargingPointType, null=True, blank=True, on_delete=models.CASCADE
+    )
 
 
 class AssocPlanProcess(models.Model):
@@ -1799,8 +1811,8 @@ class AssocAreaProcess(models.Model):
 class SimulationRange(models.Model):
     # Mutation Scenario
     scenario = models.ForeignKey(Scenario, null=False, on_delete=models.CASCADE)
-    start = models.DateTimeField(null=True)
-    end = models.DateTimeField(null=True)
+    start = models.DateTimeField(null=True, blank=True)
+    end = models.DateTimeField(null=True, blank=True)
     temperature_average = models.FloatField(
         blank=True,
         default=10,
@@ -1815,33 +1827,21 @@ class SimulationRange(models.Model):
     )
 
 
+# Unused currently but maybe reactivated in the future?
 class DepotSelection(models.Model):
     # Mutation Scenario
     scenario = models.ForeignKey(Scenario, null=False, on_delete=models.CASCADE)
     depots = models.ManyToManyField(Station)
 
 
-class ElectrificationOptions(models.Model):
-    # Mutation Scenario
-    scenario = models.ForeignKey(Scenario, null=False, on_delete=models.CASCADE)
-    gc_power_opps = models.PositiveIntegerField(
-        default=5000, null=False, validators=[MinValueValidator(1), MaxValueValidator(1000000)]
-    )
-
-    cs_power_opps = models.PositiveIntegerField(
-        default=150, null=False, validators=[MinValueValidator(1), MaxValueValidator(1000000)]
-    )
-    amount_charging_places = models.PositiveIntegerField(
-        default=2, null=False, validators=[MinValueValidator(1), MaxValueValidator(9999)]
-    )
-    station_optimization = models.BooleanField(null=False)
-    electrified_stations = models.ManyToManyField(Station)
-
-
 # Models for forms which do not mutate the scenario while in the wizard
 class VehicleTypeSelection(models.Model):
     default_vehicle_type = models.ForeignKey(
-        VehicleType, related_name="formdefaultvehicletype", null=True, on_delete=models.CASCADE
+        VehicleType,
+        related_name="formdefaultvehicletype",
+        null=True,
+        blank=True,
+        on_delete=models.CASCADE,
     )
     vehicle_type = models.ForeignKey(
         VehicleType, related_name="formvehicletype", null=False, on_delete=models.CASCADE
@@ -1849,52 +1849,30 @@ class VehicleTypeSelection(models.Model):
 
 
 class VehicleTypeMutation(models.Model):
-    scenario = models.ForeignKey(Scenario, null=True, on_delete=models.CASCADE)
+    scenario = models.ForeignKey(Scenario, null=True, blank=True, on_delete=models.CASCADE)
     original_vehicle_type = models.ForeignKey(
-        VehicleType, related_name="originalvehicletype", null=True, on_delete=models.CASCADE
+        VehicleType,
+        related_name="originalvehicletype",
+        null=True,
+        blank=True,
+        on_delete=models.CASCADE,
     )
     mutated_vehicle_type = models.ForeignKey(
-        VehicleType, related_name="mutatedvehicletype", null=True, on_delete=models.CASCADE
-    )
-
-
-class DepotMutation(models.Model):
-    scenario = models.ForeignKey(Scenario, null=True, on_delete=models.CASCADE)
-    original_depot = models.ForeignKey(
-        Depot, related_name="originaldepot", null=True, on_delete=models.CASCADE
-    )
-    mutated_original_depot = models.ForeignKey(
-        Depot, related_name="mutateddepot", null=True, on_delete=models.CASCADE
+        VehicleType,
+        related_name="mutatedvehicletype",
+        null=True,
+        blank=True,
+        on_delete=models.CASCADE,
     )
 
 
 class StationMutation(models.Model):
-    scenario = models.ForeignKey(Scenario, null=True, on_delete=models.CASCADE)
+    scenario = models.ForeignKey(Scenario, null=True, blank=True, on_delete=models.CASCADE)
     original_station = models.ForeignKey(
-        Station, related_name="originalstation", null=True, on_delete=models.CASCADE
+        Station, related_name="originalstation", null=True, blank=True, on_delete=models.CASCADE
     )
     mutated_original_station = models.ForeignKey(
-        Station, related_name="mutatedstation", null=True, on_delete=models.CASCADE
-    )
-
-
-class EnumCalculationModes(models.TextChoices):
-    AUTOMATIC = "automatic"
-    CONSTANT_POWER = "constant_power"
-    MANUAL = "manual"
-
-
-class ScenarioWizardOptions(models.Model):
-    scenario = models.ForeignKey(Scenario, null=False, on_delete=models.CASCADE)
-    station_calculation_mode = models.CharField(
-        max_length=20, choices=EnumCalculationModes.choices, null=True, default=None
-    )
-    tco_calculation_mode = models.CharField(
-        max_length=20, choices=EnumCalculationModes.choices, null=True, default=None
-    )
-
-    lca_calculation_mode = models.CharField(
-        max_length=20, choices=EnumCalculationModes.choices, null=True, default=None
+        Station, related_name="mutatedstation", null=True, blank=True, on_delete=models.CASCADE
     )
 
 
@@ -1936,9 +1914,11 @@ class AreaInformation(models.Model):
     )
     block_length = models.IntegerField(null=True, blank=True)
     vehicle_type = models.ForeignKey(VehicleType, null=False, on_delete=models.CASCADE, blank=True)
-    area_type = models.CharField(max_length=14, choices=AreaType.choices, null=True, default=None)
-    capacity = models.IntegerField(null=True)
-    power = models.FloatField(null=True)
+    area_type = models.CharField(
+        max_length=14, choices=AreaType.choices, null=True, blank=True, default=None
+    )
+    capacity = models.IntegerField(null=True, blank=True)
+    power = models.FloatField(null=True, blank=True)
 
 
 class EnumNotificationLevels(models.TextChoices):
@@ -1956,6 +1936,7 @@ class EnumNotificationType(models.TextChoices):
     MULTIPLE_DEPOT_TRIPS_IN_BLOCK_WARNING = "multi_dep_trips_in_block"
     MERGED_STATIONS_FOR_INCONSISTENT_TRIPS = "merged_station_trips_and_routes"
     INTERMEDIATE_DEPOT_STOPS_TRANSFORMED = "transformed_depot_stop_to_opp_station"
+    STATION_OPTIMIZATION_SKIPPED = "station_optimization_skipped"
     UNSTABLE_DEPOT_WARNING = "unstable_sim_w_shifting_socs"
     DELAYED_TRIP_WARNING = "delayed_trip"
     UNEXPECTED_ERROR = "unexpected_error"