+++ ebustoolbox/tasks.py
@@ -54,7 +54,6 @@ from .models import (
     copy_model_instance,
     AreaInformation,
     DepotConfigurationWish,
-    DepotMutation,
     EnumSimulationType,
     User,
     Route,
@@ -81,7 +80,6 @@ from .models import (
     UserGroup,
     SimulationRange,
     DepotSelection,
-    ElectrificationOptions,
     VehicleTypeMutation,
     VehicleTypeSelection,
     StationMutation,
@@ -777,7 +775,9 @@ def update_electrified_stations_db(electrified_stations, scenario):
         station = Station.objects.get(id=Station.get_id_from_simba_name(name), scenario=scenario)
         if not station.is_electrified:
             notification = Notification(
-                scenario=scenario,
+                # Notifications should be saved to the mutation.
+                # If the toolchain is run without a parent the notification is saved to the scenario
+                scenario=scenario.parent or scenario,
                 sender="SimBA-Optimizier from tasks.py",
                 level=EnumNotificationLevels.INFO,
                 notification_type=EnumNotificationType.ADDED_ELECTRIFICATION,
@@ -1039,7 +1039,21 @@ def run_and_merge_scenarios(
     # Run the sizing scenario with these applied changes
     assign_new_vehicles_to_db(sizing_scenario)
 
-    _ = _run_ebus_toolchain.apply(
+    if "station_optimization" not in sizing_scenario.simba_options.get("modes"):
+        Notification.objects.create(
+            # Notifications should be saved to the mutation.
+            # If the toolchain is run without a parent the notification is saved to the scenario
+            scenario=sizing_scenario.parent or sizing_scenario,
+            sender="WeBus Scenario Merge",
+            level=EnumNotificationLevels.INFO,
+            notification_type=EnumNotificationType.STATION_OPTIMIZATION_SKIPPED,
+            message=_(
+                "Die Stationsoptimierung wurde übersprungen, "
+                "da keine Stationen auf 'automatisch' gesetzt wurden."
+            ),
+        )
+
+    _run_ebus_toolchain.apply(
         (sizing_scenario_task_id,), task_id=sizing_scenario_task_id, throw=True
     )
 
@@ -1069,7 +1083,7 @@ def run_and_merge_scenarios(
     )
 
     assign_new_vehicles_to_db(average_scenario)
-    _ = _run_ebus_toolchain.apply(
+    _run_ebus_toolchain.apply(
         (default_simulation_task_id,), task_id=default_simulation_task_id, throw=True
     )
     # default_simulation_scenario = merge_scenario(mutation_id, default_simulation_task_id)
@@ -1733,12 +1747,10 @@ def deepcopy_scenario(scenario: Scenario) -> tuple[Scenario, dict]:
     """
     copied_instance, stack = core.deepcopy.deepcopy_and_sequence_reset(
         scenario,
-        exclude_models={Scenario, User, Event, Progress, UserGroup},
+        exclude_models={Scenario, User, Event, Progress, UserGroup, Notification},
         exclude_fields={
             DepotSelection._meta.get_field("depots"),
-            ElectrificationOptions._meta.get_field("electrified_stations"),
             VehicleTypeMutation._meta.get_field("original_vehicle_type"),
-            DepotMutation._meta.get_field("original_depot"),
             StationMutation._meta.get_field("original_station"),
         },
         max_depth=1,
@@ -1948,7 +1960,9 @@ def _run_ebus_toolchain(self, task_id):
                 )
             except StationOpimizationImpossible:
                 Notification.objects.create(
-                    scenario=db_scenario,
+                    # Notifications should be saved to the mutation.
+                    # If the toolchain is run without a parent the notification is saved to the scenario
+                    scenario=db_scenario.parent or db_scenario,
                     sender="SimBA-Optimizier from tasks.py",
                     level=EnumNotificationLevels.INFO,
                     notification_type=EnumNotificationType.ADDED_ELECTRIFICATION,
@@ -1959,11 +1973,13 @@ def _run_ebus_toolchain(self, task_id):
                     ),
                 )
                 schedule, args = get_schedule_from_db(db_scenario)
+            finally:
+                # Create notifications for the user
+                # if the optimizer could not achieve full electrification
+                create_negative_block_notifications(db_scenario)
         else:
             logger.info("Station optimization was skipped")
 
-        # Create notifications for the user since the optimizer could achieve full electrification
-        create_negative_block_notifications(db_scenario)
         progress.current_work += 1
         progress.status = _("Berechne das Depot")
         progress.save()
@@ -1975,7 +1991,7 @@ def _run_ebus_toolchain(self, task_id):
             logger.error("The simulation is unstable")
             logger.error(traceback.format_exception(e))
             notification = Notification(
-                scenario=db_scenario,
+                scenario=db_scenario.parent or db_scenario,
                 sender="eflips-depot",
                 level=EnumNotificationLevels.WARNING,
                 notification_type=EnumNotificationType.UNSTABLE_DEPOT_WARNING,
@@ -1992,7 +2008,7 @@ def _run_ebus_toolchain(self, task_id):
             logger.error(traceback.format_exception(e))
             # TODO: @TU what notification should the user receive
             notification = Notification(
-                scenario=db_scenario,
+                scenario=db_scenario.parent or db_scenario,
                 sender="eflips-depot",
                 level=EnumNotificationLevels.WARNING,
                 notification_type=EnumNotificationType.DELAYED_TRIP_WARNING,
@@ -2002,7 +2018,7 @@ def _run_ebus_toolchain(self, task_id):
             logger.error("Eflips raised an unexpected Exception")
             logger.error(traceback.format_exception(e))
             notification = Notification(
-                scenario=db_scenario,
+                scenario=db_scenario.parent or db_scenario,
                 sender="eflips-depot",
                 level=EnumNotificationLevels.ERROR,
                 notification_type=EnumNotificationType.UNEXPECTED_ERROR,
@@ -3153,10 +3169,12 @@ class ScheduleStationMerger:
                     scenario=scenario,
                     level=EnumNotificationLevels.WARNING,
                     notification_type=EnumNotificationType.MERGED_STATIONS_FOR_INCONSISTENT_TRIPS,
-                    message=message.format(
-                        escape(new_station.name),
-                        escape(", ".join([s.name for s in original_stations])),
-                    ),
+                    message=(
+                        message.format(
+                            escape(new_station.name),
+                            escape(", ".join([s.name for s in original_stations])),
+                        )
+                    )[:999],
                 )
 
         # After the stations were created we can change the routes
@@ -3215,7 +3233,6 @@ def transform_depot_stations(parent: Scenario, child: Scenario) -> None:
     :param parent: Source scenario
     :param child: Child scenario which is notified about changes
     """
-    reset_postgres_auto_increments(apps=[Station._meta.app_label])
     depots = Station.objects.filter(scenario=parent, charge_type=EnumChargeType.DEPOT)
     all_routes = Route.objects.filter(scenario=parent)
     depot_arrival_routes = all_routes.filter(arrival_station__in=depots)
@@ -3334,7 +3351,7 @@ def transform_depot_stations(parent: Scenario, child: Scenario) -> None:
                     f"Für den Umlauf {escape(rotation.name)} wurden Zwischenhaltestellen "
                     f"an den Depots {[escape(s.name) for s in stations]} erzeugt. "
                     "Mehr Informationen finden Sie in der Hilfe."
-                ),
+                )[:999],
             )
     if len(changed_rotations) > 0:
         logger.warning(
@@ -3510,18 +3527,40 @@ def create_negative_block_notifications(scenario: Scenario) -> None:
     ).select_related("trip__rotation")
     if not events.exists():
         return
-    low_soc_blocks = {event.trip.rotation.name for event in events}
+    low_soc_blocks = {str(event.trip.rotation.name) for event in events}
     Notification.objects.create(
         scenario=scenario,
         sender="SimBA-Optimizier from tasks.py",
         level=EnumNotificationLevels.INFO,
         notification_type=EnumNotificationType.LOW_SOC_BLOCKS,
         message=_(
-            "Die Stationsoptimierung konnte nicht alle Umläufe elektrifzieren. "
-            f"Folgende {len(low_soc_blocks)} Umläufe haben auch nach der Optimierung einen SOC unter 0%: "
-            + ", ".join(low_soc_blocks)
+            (
+                "Die Stationsoptimierung konnte nicht alle Umläufe elektrifzieren. "
+                f"Folgende {len(low_soc_blocks)} Umläufe haben auch nach der Optimierung einen SOC unter 0%: "
+                + ", ".join(low_soc_blocks)
+            )[:999]
         ),
     )
+
+
+def delete_scenario(scenario: Scenario):
+    """Delete the scenario and appropriate relatives
+
+    In case of a SimulationScenario only this scenario is deleted.
+    In case of a MutationScenario the parent is deleted if the parent has no other children.
+    This is done since the user does not see source scenarios inside his management view.
+    """
+    logger.info(f"Deleting Scenario with {scenario.id=} and {scenario.manager}")
+    # Delete from db. This does not affect the in memory scenario
+    logger.info(scenario.delete())
+    if scenario.scenario_type == EnumScenarioType.SIMULATION:
+        return
+    if scenario.parent is None:
+        return
+    children = Scenario.objects.filter(parent=scenario.parent)
+    if children.exists():
+        return
+    logger.info(scenario.parent.delete())
 
 
 class StationOpimizationImpossible(Exception):